Lexer extensions for the 'for' loop
Now that we know how to add basic control flow constructs to the
language, we have the tools to add more powerful things. Lets add
something more aggressive, a ‘for’ expression:

extern putchard(char)
def printstar(n)
  for i = 1, i < n, 1.0 in
    putchard(42);  # ascii 42 = '*'

printstar(100);

This expression defines a new variable (“i” in this case) which iterates
from a starting value, while the condition (“i < n” in this case) is
true, incrementing by an optional step value (“1.0” in this case). If
the step value is omitted, it defaults to 1.0. While the loop is true,
it executes its body expression. Because we don’t have anything better
to return, we’ll just define the loop as always returning 0.0. In the
future when we have mutable variables, it will get more useful.

As before, lets talk about the changes that we need to Kaleidoscope to
support this.

The lexer extensions are the same sort of thing as for if/then/else.