Add PHI for 'for' loop
Once we have that, we create the actual block that starts the loop and
create an unconditional branch for the fall-through between the two
blocks.

[code]

Now that the “preheader” for the loop is set up, we switch to emitting
code for the loop body. To begin with, we move the insertion point and
create the PHI node for the loop induction variable. Since we already
know the incoming value for the starting value, we add it to the Phi
node. Note that the Phi will eventually get a second value for the
backedge, but we can’t set it up yet (because it doesn’t exist!).