Save old value of loop iter; emit 'for' loop body
Now the code starts to get more interesting. Our ‘for’ loop introduces a
new variable to the symbol table. This means that our symbol table can
now contain either function arguments or loop variables. To handle this,
before we codegen the body of the loop, we add the loop variable as the
current value for its name. Note that it is possible that there is a
variable of the same name in the outer scope. It would be easy to make
this an error (emit an error and return null if there is already an
entry for VarName) but we choose to allow shadowing of variables. In
order to handle this correctly, we remember the Value that we are
potentially shadowing in OldVal (which will be null if there is no
shadowed variable).

Once the loop variable is set into the symbol table, the code
recursively codegen’s the body. This allows the body to use the loop
variable: any references to it will naturally find it in the symbol
table.