Finish 'for' loop codegen
The final code handles various cleanups: now that we have the “NextVar”
value, we can add the incoming value to the loop PHI node. After that,
we remove the loop variable from the symbol table, so that it isn’t in
scope after the for loop. Finally, code generation of the for loop
always returns 0.0, so that is what we return from ForExprAST::Codegen.