Extend PrototypeAST to accommodate operators
On the other hand, we have to be able to represent the definitions of
these new operators, in the “def binary| 5” part of the function
definition. In our grammar so far, the “name” for the function
definition is parsed as the “prototype” production and into the
PrototypeAST AST node. To represent our new user-defined operators as
prototypes, we have to extend the PrototypeAST AST node like this:

[code]

Basically, in addition to knowing a name for the prototype, we now keep
track of whether it was an operator, and if it was, what precedence
level the operator is at. The precedence is only used for binary
operators (as you’ll see below, it just doesn’t apply for unary
operators).