Extend ParsePrototype to accommodate operators
Now that we have a way to represent the prototype for a user-defined
operator, we need to parse it:

[code]

This is all fairly straightforward parsing code, and we have already
seen a lot of similar code in the past. One interesting part about the
code above is the couple lines that set up FnName for binary operators.
This builds names like “binary@” for a newly defined “@” operator. This
then takes advantage of the fact that symbol names in the LLVM symbol
table are allowed to have any character in them, including embedded null
characters.