Install operator precedence upon definition
The final piece of code we are missing, is a bit of top-level magic:

[code]

Basically, before codegening a function, if it is a user-defined
operator, we register it in the precedence table. This allows the binary
operator parsing logic we already have in place to handle it. Since we
are working on a fully-general operator precedence parser, this is all
we need to do to “extend the grammar”.

Now we have useful user-defined binary operators. This builds a lot on
the previous framework we built for other operators. Adding unary
operators is a bit more challenging, because we don’t have any framework
for it yet - lets see what it takes.