Add class for unary op expressions
Since we don’t currently support unary operators in the Kaleidoscope
language, we’ll need to add everything to support them. Above, we added
simple support for the ‘unary’ keyword to the lexer. In addition to
that, we need an AST node:

[code]

This AST node is very simple and obvious by now. It directly mirrors the
binary operator AST node, except that it only has one child.