Parse unary operator invokations
With this, we need to add the parsing logic. Parsing a unary operator is
pretty simple: we’ll add a new function to do it:

[code]

The grammar we add is pretty straightforward here. If we see a unary
operator when parsing a primary operator, we eat the operator as a
prefix and parse the remaining piece as another unary operator. This
allows us to handle multiple unary operators (e.g. ”!!x”). Note that
unary operators can’t have ambiguous parses like binary operators can,
so there is no need for precedence information.