Add parser support for unary op definitions
With these two simple changes, we are now able to parse unary operators
and build the AST for them. Next up, we need to add parser support for
prototypes, to parse the unary operator prototype. We extend the binary
operator code above with:

[code]

As with binary operators, we name unary operators with a name that
includes the operator character. This assists us at code generation
time.