Start refactoring to accommodate variable mutation
The symbol table in Kaleidoscope is managed at code generation time by
the ‘NamedValues‘ map. This map currently keeps track of the LLVM
“Value*” that holds the double value for the named variable. In order to
support mutation, we need to change this slightly, so that it
NamedValues holds the memory location of the variable in question. Note
that this change is a refactoring: it changes the structure of the code,
but does not (by itself) change the behavior of the compiler. All of
these changes are isolated in the Kaleidoscope code generator.

At this point in Kaleidoscope’s development, it only supports variables
for two things: incoming arguments to functions and the induction
variable of ‘for’ loops. For consistency, we’ll allow mutation of these
variables in addition to other user-defined variables. This means that
these will both need memory locations.

To start our transformation of Kaleidoscope, we’ll change the
NamedValues map so that it maps to AllocaInst* instead of Value*. Once
we do this, the C++ compiler will tell us what parts of the code we need
to update.

Also, since we will need to create these alloca’s, we’ll use a helper
function that ensures that the allocas are created in the entry block of
the function:

[code]

This funny looking code creates an IRBuilder object that is pointing at
the first instruction (.begin()) of the entry block. It then creates an
alloca with the expected name and returns it. Because all values in
Kaleidoscope are doubles, there is no need to pass in a type to use.