Update ForExprAST::Codegen
As you can see, this is pretty straightforward. Now we need to update
the things that define the variables to set up the alloca. We’ll start
with ForExprAST::Codegen (see the full code listing for the unabridged
code):

[code]

This code is virtually identical to the code before we allowed mutable
variables. The big difference is that we no longer have to construct a
PHI node, and we use load/store to access the variable as needed.