Add precedence for the assignment operator
Now that all symbol table references are updated to use stack variables,
we’ll add the assignment operator.

With our current framework, adding a new assignment operator is really
simple.  We will parse it just like any other binary operator, but
handle it internally (instead of allowing the user to define it). The
first step is to set a precedence.