Start codegen for assignment
Now that the parser knows the precedence of the binary operator, it
takes care of all the parsing and AST generation. We just need to
implement codegen for the assignment operator. This looks like:

[code]

Unlike the rest of the binary operators, our assignment operator doesn’t
follow the “emit LHS, emit RHS, do computation” model. As such, it is
handled as a special case before the other binary operators are handled.
The other strange thing is that it requires the LHS to be a variable. It
is invalid to have “(x+1) = expr” - only things like “x = expr” are
allowed.