Add token defs for variable definitions
Now that we have an assignment operator, we can mutate loop variables
and arguments. For example, we can now run code like this:

extern printd(x);

def binary : 1 (x y) y;

def test(x)
  printd(x) :
  x = 4 :
  printd(x);

test(123);
When run, this example prints “123” and then “4”, showing that we did
actually mutate the value! Okay, we have now officially implemented our
goal: getting this to work requires SSA construction in the general
case. However, to be really useful, we want the ability to define our
own local variables, lets add this next!

Adding var/in is just like any other other extensions we made to
Kaleidoscope: we extend the lexer, the parser, the AST and the code
generator. The first step for adding our new ‘var/in’ construct is to
extend the lexer. As before, this is pretty trivial, the code looks like
this:

[code]