Add VarExprAST to represent a variable declaration
The next step is to define the AST node that we will construct. For
var/in, it looks like this:

[code]

var/in allows a list of names to be defined all at once, and each name
can optionally have an initializer value. As such, we capture this
information in the VarNames vector. Also, var/in has a body, this body
is allowed to access the variables defined by the var/in.