Start codegen for VarExprAST
Now that we can parse and represent the code, we need to support
emission of LLVM IR for it. This code starts out with:

[code]

Basically it loops over all the variables, installing them one at a
time. For each variable we put into the symbol table, we remember the
previous value that we replace in OldBindings.