Define DebugInfo struct to hold debug info
The top level container for a section of code in DWARF is a compile
unit. This contains the type and function data for an individual
translation unit (read: one file of source code). So the first thing we
need to do is construct one for our fib.ks file.

Similar to the IRBuilder class we have a DIBuilder class that helps in
constructing debug metadata for an llvm IR file. It corresponds 1:1
similarly to IRBuilder and llvm IR, but with nicer names. Using it does
require that you be more familiar with DWARF terminology than you needed
to be with IRBuilder and Instruction names, but if you read through the
general documentation on the Metadata Format it should be a little more
clear. We’ll be using this class to construct all of our IR level
descriptions. Construction for it takes a module so we need to construct
it shortly after we construct our module. We’ve left it as a global
static variable to make it a bit easier to use.

Next we’re going to create a small container to cache some of our
frequent data.  The first will be our compile unit, but we’ll also write
a bit of code for our one type since we won’t have to worry about
multiple typed expressions:

[code]