Create DIBuilder object in main
There are a couple of things to note here. First, while we’re producing a
compile unit for a language called Kaleidoscope we used the language constant
for C. This is because a debugger wouldn’t necessarily understand the calling
conventions or default ABI for a language it doesn’t recognize and we follow the
C ABI in our llvm code generation so it’s the closest thing to accurate. This
ensures we can actually call functions from the debugger and have them execute.
Secondly, you’ll see the “fib.ks” in the call to createCompileUnit. This is a
default hard coded value since we’re using shell redirection to put our source
into the Kaleidoscope compiler. In a usual front end you’d have an input file
name and it would go there.