Create Compile Unit in PrototypeAST::Codegen
Now that we have our Compile Unit and our source locations, we can add
function definitions to the debug info. So in PrototypeAST::Codegen we
add a few lines of code to describe a context for our subprogram, in
this case the “File”, and the actual definition of the function itself.

So the context:

[code]

giving us a DIFile and asking the Compile Unit we created above for the
directory and filename where we are currently.